// src/models/seed-metadata.model.ts
export interface SeedMetadata {
  seed: string;
  item_placement: string;
  dungeon_items: string;
  accessibility: string;
  goal: string;
  entry_crystals_tower: string;
  entry_crystals_ganon: string;
  mode: string;
  enemizer: string;
  weapons: string;
  mmMedallion: string;
  trMedallion: string;
}